refactor: moved Widgets component back inside Dashboard route
